requirements.txt
torch
transformers
pytube
librosa
opencv-python
Pillow
speechrecognition
ffmpeg
